+++ src/components/post/PhotoList.js
@@ -1,15 +1,14 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { ImageBlock, ImageSrc } from './PhotoList.styles';
+import PostCarousel from './PostCarousel';
 
 function PhotoList({ image }) {
   const { id, alt_description } = image;
   return (
     <div>
       test중
-      <ImageBlock>
-        <ImageBlock key={id}>
-          <ImageSrc alt={alt_description} src={image.urls.small} />
-        </ImageBlock>
+      <ImageBlock key={id}>
+        <ImageSrc alt={alt_description} src={image.urls.small} />
       </ImageBlock>
     </div>
   );
