+++ src/components/post/Posts.js
@@ -3,7 +3,6 @@ import Cards from '../common/Cards';
 import PostHeader from './PostHeader';
 import PostContents from './PostContents';
 import PostCommentList from './PostCommentList';
-import Test from './test';
 import PostUpload from './PostUpload';
 
 function Posts() {
